Update tailwind.config.js and index.css styling (Light theme fix)
The base-100 color in the tailwind.config.js has been updated, and redundant color and background-color properties in index.css have been removed. Additionally, in the btn-purple class, border properties are now set to 0, and the media color scheme for light is completely removed for a cleaner and more uniform look in index.css.